Sólo los usuarios admin pueden borrar otros usuarios. Se añade en la tabla usuarios campo admin, por ahora sólo se crea un admin de forma manual.

+++ src/main/java/com/s2daw/demo/controllers/UsuarioController.java
@@ -46,12 +46,19 @@ public class UsuarioController {
     public void eliminarUsuario(@RequestHeader(value="Authorization") String token,
                                 @PathVariable Long id){
         if (!validarToken(token)) return;
+        if (!soyAdmin(token)) return;
         usuarioDao.eliminarUsuario(id);
+        // Lo suyo es contestar si va todo bien o no
     }
 
     // función de apoyo
     private boolean validarToken(String token){
         String usuarioid=jwtUtil.getKey(token);
         return usuarioid!=null;
+    }
+
+    private boolean soyAdmin(String token){
+        Boolean admin=jwtUtil.getAdmin(token);
+        return admin;
     }
 }
